push a droplet instead of a package
this avoids the extra time and resources it takes to stage an app, and
it works for our use case because we aren't testing the stability of the
system during the process of pushing N apps.

it's also worth noting that this specific app requires a minimum of 16M
of memory and 75M of disk to run successfully without any quota issues.
this should be altered if the droplet changes.